Move this somewhere it can more easily be reused
